% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctasval.R
\name{anomaly_sd}
\alias{anomaly_sd}
\title{Add Standard Deviation Anomalies}
\usage{
anomaly_sd(df, anomaly_degree, site = "sample_site")
}
\arguments{
\item{df}{Data frame containing the study data.}

\item{anomaly_degree}{Degree of anomaly to add.}

\item{site}{The site to sample from. Default is "sample_site".}
}
\value{
A data frame with added anomalies.
}
\description{
This function adds standard deviation anomalies to the \code{result} column of the data frame.
}
