% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly.R
\name{cumsum_fraction}
\alias{cumsum_fraction}
\title{Add a fraction of lag to vector}
\usage{
cumsum_fraction(x, fraction)
}
\description{
Add a fraction of lag to vector
}
\examples{
\dontrun{
cumsum_fraction(rnorm(10), 0.5)
cumsum_fraction(rnorm(1), 0.1)
}
}
\keyword{internal}
