% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctasval.R
\name{get_anomaly_data}
\alias{get_anomaly_data}
\title{Generate Anomaly Data}
\usage{
get_anomaly_data(
  df,
  n_sites,
  fun_anomaly,
  anomaly_degree,
  site_prefix = "site"
)
}
\arguments{
\item{df}{Data frame containing the study data.}

\item{n_sites}{Number of sites to generate.}

\item{fun_anomaly}{Function to apply to generate anomalies.}

\item{anomaly_degree}{Degree of anomaly to add.}

\item{site_prefix}{Prefix for the site names. Default is "site".}
}
\value{
A data frame with the generated anomaly data.
}
\description{
This function generates anomaly data by applying a specified anomaly function to the data frame.
}
\seealso{
\code{\link{get_anomaly_scores}}
}
\keyword{internal}
