% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctasval.R
\name{get_ctas}
\alias{get_ctas}
\title{Get CTAS}
\usage{
get_ctas(
  df,
  feats,
  default_minimum_timepoints_per_series = 3,
  default_minimum_subjects_per_series = 3,
  default_max_share_missing_timepoints_per_series = 0.5,
  default_generate_change_from_baseline = FALSE,
  autogenerate_timeseries = TRUE,
  site_scoring_method
)
}
\arguments{
\item{df}{Data frame containing the study data.}

\item{feats}{Features to calculate for the timeseries.}

\item{default_minimum_timepoints_per_series}{Minimum timepoints per series. Default is 3.}

\item{default_minimum_subjects_per_series}{Minimum subjects per series. Default is 3.}

\item{default_max_share_missing_timepoints_per_series}{Maximum share of missing timepoints per series. Default is 0.5.}

\item{default_generate_change_from_baseline}{Logical indicating whether to generate change from baseline. Default is FALSE.}

\item{autogenerate_timeseries}{Logical indicating whether to auto-generate timeseries. Default is TRUE.}
}
\value{
A data frame with the CTAS results.
}
\description{
This function processes a study using the CTAS (Clinical Trial Anomaly Spotter) by providing various parameters and features.
}
\seealso{
\code{\link{get_anomaly_scores}}
}
\keyword{internal}
