% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctasval.R
\name{prep_sdtm}
\alias{prep_sdtm}
\title{Prepare SDTM Data}
\usage{
prep_sdtm(lb, vs, dm, scramble = TRUE)
}
\arguments{
\item{lb}{Data frame containing the LB data.}

\item{vs}{Data frame containing the VS data.}

\item{dm}{Data frame containing the DM data.}

\item{scramble}{Logical indicating whether to scramble the SITEID in the DM data. Default is TRUE.}
}
\description{
Prepare SDTM Data
}
