% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctasval.R
\name{prep_sdtm_lb}
\alias{prep_sdtm_lb}
\title{Prepare SDTM LB Data}
\usage{
prep_sdtm_lb(lb, dm, scramble = TRUE)
}
\arguments{
\item{lb}{Data frame containing the LB data.}

\item{dm}{Data frame containing the DM data.}

\item{scramble}{Logical indicating whether to scramble the SITEID in the DM data. Default is TRUE.}
}
\value{
A data frame with the prepared SDTM LB data.
}
\description{
This function prepares the LB (Laboratory) data for SDTM (Study Data Tabulation Model) by merging it with the DM (Demographics) data.
}
