% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctasval.R
\name{prep_sdtm_vs}
\alias{prep_sdtm_vs}
\title{Prepare SDTM VS Data}
\usage{
prep_sdtm_vs(vs, dm, scramble = TRUE)
}
\arguments{
\item{vs}{Data frame containing the VS data.}

\item{dm}{Data frame containing the DM data.}

\item{scramble}{Logical indicating whether to scramble the SITEID in the DM data. Default is TRUE.}
}
\value{
A data frame with the prepared SDTM LB data.
}
\description{
This function prepares the VS (Vital Sign) data for SDTM (Study Data Tabulation Model) by merging it with the DM (Demographics) data.
}
