% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly.R
\name{rsurprise}
\alias{rsurprise}
\title{sample values from a randomly selected non-normal distribution}
\usage{
rsurprise(n, verbose = FALSE)
}
\arguments{
\item{n}{integer, number of values}

\item{verbose}{logical, Default: FALSE}
}
\description{
sample values from a randomly selected non-normal distribution
}
\examples{
set.seed(1)
rsurprise(5, verbose = TRUE)
rsurprise(5, verbose = TRUE)


}
