% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctasval.R
\name{sample_site}
\alias{sample_site}
\title{Sample Site Data}
\usage{
sample_site(df, site = "sample_site")
}
\arguments{
\item{df}{Data frame containing the study data.}

\item{site}{The site to sample from. Default is "sample_site".}
}
\value{
A data frame with the sampled site data.
}
\description{
This function samples one random site per parameter from the given data frame.
}
\seealso{
\code{\link{anomaly_average}}, \code{\link{anomaly_sd}}
}
\keyword{internal}
