% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctasval.R
\name{get_anomaly_scores}
\alias{get_anomaly_scores}
\title{Get Anomaly Scores}
\usage{
get_anomaly_scores(
  df,
  n_sites,
  fun_anomaly,
  anomaly_degree,
  feats,
  thresh = NULL,
  default_minimum_timepoints_per_series = 3,
  default_minimum_subjects_per_series = 3,
  default_max_share_missing_timepoints_per_series = 0.5,
  default_generate_change_from_baseline = FALSE,
  autogenerate_timeseries = TRUE,
  site_scoring_method,
  padjust_method = "BY"
)
}
\arguments{
\item{df}{Data frame containing the study data.}

\item{n_sites}{Number of sites to generate.}

\item{fun_anomaly}{Function to apply to generate anomalies.}

\item{anomaly_degree}{Degree of anomaly to add.}

\item{feats}{Features to calculate for the timeseries.}

\item{thresh}{Threshold for classification. Default is NULL.}

\item{default_minimum_timepoints_per_series}{Minimum timepoints per series. Default is 3.}

\item{default_minimum_subjects_per_series}{Minimum subjects per series. Default is 3.}

\item{default_max_share_missing_timepoints_per_series}{Maximum share of missing timepoints per series. Default is 0.5.}

\item{default_generate_change_from_baseline}{Logical indicating whether to generate change from baseline. Default is FALSE.}

\item{autogenerate_timeseries}{Logical indicating whether to auto-generate timeseries. Default is TRUE.}

\item{site_scoring_method}{How to score sites ("ks" = Kolmogorov-Smirnov, "mixedeffects" = mixed effects modelling, "avg_feat_value" = Average site feature value.}

\item{padjust_method}{parameter passed to \code{\link[=p.adjust]{p.adjust()}} method parameter, Default: "BY"}
}
\value{
A data frame with the anomaly scores.
}
\description{
This function calculates the scores for anomalies in the data frame.
}
\seealso{
\code{\link{ctasval}}
}
\keyword{internal}
